R183: regression tests, the framing module split out, and the round's documentation
TESTS — tests/r183-checks.test.mjs (13). The three defects this round fixed had the same
shape: something failed, nothing said so, and the UI kept rendering a plausible value. A test
that only asserts the happy path would have passed on every one of them, so these aim at the
failure modes — a 429 whose error body is valid JSON, a sunrise that answers for the wrong
DAY, a bounding box that is a synthetic node stub or spans the globe, a counter that started
counting parts instead of aircraft. Sunrise is pinned against MET Norway's Sunrise 3.0 values;
the zoom table is asserted by SHAPE (continent < country < city < station < building).

js/place-framing.js (new) — the framing decision moved out of js/search-geocode.js. Three
existing invariants said it did not belong there, and each one is worth stating because they
are exactly the guard rails earlier rounds installed:
  · r169 #4 — a split factory's body may contain only DECLARATIONS, and publishing a global
    from inside one is a statement that RUNS;
  · r169 #2 — a factory's return list is the app-body shim contract, pinned name-for-name;
  · r175 — src/main.js's import order is pinned, so wx-source sits after newsgeo rather than
    weakening "newsgeo is first among the feature modules".
Being map-free is also what lets the whole table be tested without a browser, which is how
three of its four traps were found in the first place.

Getting there took a wrong cut worth recording: the first extraction sliced from the R183
banner to `gotoPlace` and swallowed doGeocode and the whole result-card block with it. The
split-scope checker caught it immediately and by name ("free identifier doGeocode resolves to
nothing at runtime") — the #R163 parser-backed check doing exactly the job it was built for.

DOCS — Architecture.md gains js/wx-source.js and the note that weather.js delegates to it;
DEV-NOTES.md gets the R183 entry, which states plainly that 3 of the 11 requested items are
done and 7 are not, rather than implying a complete round.
