+++ DEV-NOTES.md
@@ -23,6 +23,162 @@
 
 ---
 
+## R183 — **「更新されない」の正体は他人のクォータ／検索のズームを対象に合わせる／航空機を見えるようにして立体に骨を入れる** (tag `#R183`)
+
+指示は11件（ウィジェット・衛星画像/3Dの高速高画質化・Cesiumのフライトシム・人工衛星レイヤー・
+航空機アイコン・地名検索のズーム・3D Volumeの複数オブジェクト・Line of sight・ドローン航法・経路検索）。
+**このラウンドで完了したのは先頭3件**（下記）。残り7件は未着手で、#R184 以降に持ち越す。
+「途中まで作って全部入りに見せる」ことはしない——標準指示4。
+
+---
+
+### 1. UV Index と weather のウィジェットが更新されない
+
+**まず再現した。** アプリ自身のオリジンから：
+
+| エンドポイント | 応答 |
+|---|---|
+| `api.open-meteo.com/v1/forecast` | **429** `{"error":true,"reason":"Daily API request limit exceeded. Please try again tomorrow."}` |
+| `air-quality-api.open-meteo.com/v1/air-quality` | 200（**別クォータ**） |
+| `api.met.no/…/locationforecast/2.0/complete` | 200・CORS可・キー不要 |
+
+報告された組み合わせと完全に一致する。`weather` と `uv` は forecast ホストを読み、`aqi` は
+air-quality ホストを読む——だから前の2つだけが死んでいた。Open-Meteo の無料枠は **IP 単位かつ
+製品ホスト単位**で、このアプリは forecast ホストの重い消費者（36箇所・ウィジェット板だけで3分ごとに
+weather + uv + sun）。
+
+**そして失敗が無音だった。** `js/widgets.js` は `r.ok` を見ずに `await r.json()` していた。
+エラーボディも**正しいJSON**なので、`j.current` が undefined になり、カードは
+
+```
+🌤 —          ← 永久に。理由はどこにも出ない
+```
+
+を描き続ける。外から見るとこれが「更新されない」である。
+
+**#R72 は同じ欠陥を地点天気ポップアップで直していた**（キャッシュ・MET Norway フォールバック・
+正直なエラー行）。その修正は**旅をしなかった**——旅をする通路が無かったからだ。呼び出し側が各自
+`fetch(...).then(r=>r.json())` を書いていた。11ラウンド後もウィジェット板に同じバグが残っていた。
+
+→ `js/wx-source.js`（新規）= `window.IntMapWx`、**唯一の取得口**。
+`js/weather.js` は動作を変えずに委譲する（MET マッピングの二重持ちが、そもそも修正が渡らなかった原因）。
+
+**「もう一方を試す」より一段深い部分：**
+- **サーキットブレーカー**。1日上限の429は「今日いっぱいそうである」という宣言であって、再要求は
+  無駄なだけでなく**クォータを0のままにする当の原因**。localStorage に記録し、00:00 UTC まで
+  fallback 直行。それ以外の失敗は10分バックオフ（そちらは本当に一過性）。
+- **合流＋TTLキャッシュ**（URL単位）。同じ座標の5ウィジェットが3分ごとに5リクエスト → 1本。
+- **日の出・日没は計算に**。純粋な天文なのに forecast ホストを叩いていたので、クォータ切れで
+  一緒に死んでいた。ネットワーク非依存になった。
+
+**単位の正直さ（数値の意味が出所で変わった2件）：**
+- Open-Meteo の `uv_index` は**全天**、MET の `ultraviolet_index_clear_sky` は雲を無視する**上限**＝
+  別の量。同じフィールドに入れず、カードは「快晴時 / clear sky」と明示する。
+- MET の系列は**現在時刻から先**しか無いので初日バケットは「今日の残り」。`_partialFirstDay` を立て、
+  カードは「本日最大」ではなく「この先の最大 / peak ahead」と言う。
+
+**自分の出力を読んで見つけた3件**（レビューではなく実測）：
+- **日の出が1日ずれていた。** 日の出方程式は「ある瞬間に最も近い**太陽日**」を答え、ウィジェットが
+  訊かれているのは「利用者の時計の上の**暦日**」。両者は真夜中付近でずれ、**経度と時間帯が一致しない
+  国では数時間ずれる**（アイスランドは−22°でUTC、アラスカは−150°で−8）。「今」を基準にすると東京が
+  前日を answer した。**その暦日の地方平均太陽正午**を基準にして解決。MET の Sunrise 3.0 API を正解として
+  10都市×7日＝**70点で誤日0・最悪誤差4.7分**（レイキャヴィーク、白夜で地平線をかすめる分）。
+- **MET のフォールバックが系列を UTC 日でバケットしていた。** 08:00 JST では「今日」に1時間しか
+  入らず、**UV最大が現在値と同じ2.9**、天気カードが **H 32 / L 32**。地方日でバケットして
+  UV最大 8.6・**H 37 / L 28**。
+- **「13h 60m」**。分に丸めてから時分に割る。
+
+---
+
+### 2. 地名検索のズームを対象に合わせる
+
+`gotoPlace` は**何を検索しても `zoom:9`**（約100km幅）だった。建物・駅・公園には遠すぎ、国・海・砂漠には
+近すぎる。z9 が正解になる対象はほとんど無い。
+
+**2段構え（片方は実測、もう片方は推定なので、その順）：**
+1. **ジオコーダが公開している実際の範囲** — Nominatim の `boundingbox`、Photon の `extent`。
+   推定ではなく対象自身の footprint なので、バチカンも日本も画面いっぱいになる。
+   **Photon の `extent` と `osm_key`/`osm_value` は捨てられていた**ので全ヒットが既定値に落ちていた。
+2. **範囲が使えないときは種別** — Nominatim の category/type/**addresstype**、Photon の
+   osm_key/osm_value、Open-Meteo の GeoNames `feature_code`、自前 gazetteer の `kind`
+   （`gotoPlace` に null で届いていたので通した）。
+
+**罠4件。すべて実応答を流して見つけた（推論ではない）：**
+- **ロシア・バチカン・東京は category/type が全部同じ**（boundary/administrative）。区別するのは
+  `addresstype`（country / country / province）。ただし**行政区画と集落に限って**先に読む——
+  サハラ砂漠の addresstype は "region" で、4,800kmの砂漠が郡のように枠取りされた。物理地物は
+  category/type（natural/desert）が正確なのでそちらが勝つ。
+- **OSMで単一ノードの地物には footprint が無く**、Nominatim は **±0.0001°（約11m）の合成箱**を返す。
+  富士山・サハラ・太平洋・渋谷の地下鉄入口はすべてノード。この箱で枠取りすると
+  **サハラが zoom 16.5**＝砂を11m見る。両軸 0.0005° 未満の箱は無視。
+  （アフリカもノードだが**意図的な50°×50°の箱**を持っていて生き残る——規則はノードではなく合成箱を見る。）
+- **日付変更線をまたぐ範囲は経度360°で返る**（ロシア、実測）。棄却するだけでは足りない：
+  **棄却されたこと自体が「巨大な地物である」証拠**なので、既定の country ズーム4.4（シベリアの中に
+  着地して四辺から国がはみ出す）ではなく z3.2 に広げる。
+- **行政範囲は飛び地を含む。** 東京都は小笠原まで南に1,000km あり、実範囲18.4°×15.7°で枠取りすると
+  **本州の南の海のど真ん中が中心**になり都心は上端の点になる——幾何としては正しく、答えとしては無用。
+  見分け方は「**ジオコーダ自身の点が箱の中心付近に無い**」。両軸の中央60%から外れる範囲は信用しない。
+  コンパクトな対象（日本・バイカル湖・セントラルパーク）は自分の範囲を保つ。
+
+**実クエリ18件の実測（全部が以前は zoom 9）：**
+ロシア3.2・アフリカ3.2・フランス3.2・日本4.0・サハラ4.6・バイカル湖6.0・東京6.2（旧4.7＝海の上）・
+アマゾン川9.0・京都9.8・マンハッタン10.8・富士山11.0・ヒースロー13.2・セントラルパーク13.3・
+渋谷駅14.6・ダウニング街10番15.6・バチカン15.6・エッフェル塔16.5・太平洋2.6
+
+---
+
+### 3. 航空機アイコンを目立たせ、立体に骨を入れる
+
+**平面グリフが薄かった理由は3つあり、大きさはそのうち1つだけだった：**
+- **2倍画面に1倍で焼いていた。** `addImage` に 44px の ImageData を `pixelRatio` 無しで渡していたので、
+  MapLibre は 44 canvas px を 44 CSS px として扱い、HiDPI では GPU が拡大する——**にじんだ飛行機**。
+  #R178/#R179/#R180 の半解像度ラスタと**同じ系統の欠陥**＝入口で画素を捨てている。
+  devicePixelRatio で描き、その比を宣言する。**同じ画面サイズのまま鮮明**になる（大きくはならない）。
+- **地図から分離していなかった。** 平塗り＋1.6pxの白線は、明るい衛星画像・雪・淡い基図の上で消える。
+  下に柔らかいドロップシャドウ、上に太い白リムを置き、**背景に依存せず自前でコントラストを持つ**。
+- **形が紙飛行機だった。** 直線の翼・エンジン無し・水平尾翼が独立していない。後退翼・ナセル2基・
+  絞った機首・独立した水平尾翼の**実機の平面形**に。30pxでも「矢印」ではなく「航空機」に見える。
+サイズ勾配も上げた（z2: 0.4→0.52、z9: 0.78→1.0）。
+
+**立体。** #R172 は実高度を与えた——重要な半分だ。しかし置かれていたのは**平面形を一様な厚みで
+押し出した1枚**で、傾けると空中に置かれた紙の切り抜きに見えた。翼と垂直尾翼が同じ高さにある
+飛行機は無い。`fill-extrusion` は MapLibre 5.24 で実高度を持つ唯一のプリミティブのままだが、
+**胴体が1つの押し出しである必要はどこにも無かった**。4つにした：胴体（最も高い実体・全長）・
+主翼（胴体の低い位置の薄い板、ナセルを吊る）・水平尾翼（翼線より上の小さい板）・
+垂直尾翼（すべての上に立つ細い板）。真上からの輪郭は不変、傾けると骨格が出る。
+
+**そして、サイズの門を間違った軸に掛けていた（記録する価値がある）。** 最初の版は
+**画面上のグリフサイズ**で部品を出し分けた。もっともらしいが、ここでは無意味である——`half` は
+**どのズームでも26pxの下限にクランプ**されており（z4でも航空機が見えるのはそのおかげ）、
+z15あたりまで `half/mpp` は13pxに固定される。**部品は一度も出なかった。**
+実測（東京 z13・実トラフィック）：**53機・91フィーチャ ＝ 53シルエット＋38ポスト・部品0**。
+実際に変動する量は**機数**なので、予算をそちらに掛ける（≤900機／フィードの上限は1800）。
+z11で再測：**57機 → 273フィーチャ ＝ 57×4 ＋ 45ポスト**。
+
+**数え間違いも1件。** `lifted` は「地面から離れたフィーチャ数」で、これが「飛行中の機数」と一致するのは
+1機＝1フィーチャの間だけだった。**4倍を報告するところ**だった。機体は胴体で数え、生のフィーチャ数は
+自分の名前を持つ（#R181「カウンタが数えている物を疑え」）。
+
+**自分で踏んだ TDZ も1件。** 共有する平面形定数を最初 `ensurePlaneIcons` の隣＝**最初の使用より
+約150行下**に書いた。同一スコープの `const` なので**自分自身の temporal dead zone**に入り、
+ファクトリ構築中に ReferenceError を投げて **data-layers モジュールごと落とす**。#R167 の通り、
+全使用箇所より上で宣言する。
+
+---
+
+### 4. テスト
+
+`tests/r183-checks.test.mjs`（13件）。このラウンドの3欠陥は**同じ形**をしていた——何かが失敗し、
+誰も何も言わず、UIはもっともらしい値を描き続ける。「正常系が通ること」だけを見るテストは
+**3件とも素通りさせる**。だから狙いは**失敗の側**に置いた：
+HTTP 429 の**正しいJSONエラーボディ**／**日を間違える**日の出計算／合成スタブ箱と地球を跨ぐ箱／
+部品を数え始めたカウンタ。
+
+日の出は MET Sunrise 3.0 の実値を正解として持ち、ズーム表は**形**を主張する
+（大陸 < 国 < 都市 < 駅 < 建物）。
+
+---
+
 ## R182 — **Cesium の「操作がほぼできない」を、MapLibre のジェスチャそのものを転写して直す／ついでに出てきた、動く視点の着地点が外れていた件** (tag `#R182`)
 
 指示は3件。1件目「MabLibre依存脱却作業を完了させて」、2件目「デフォルトはこれまで通りMapLibreだが、
