+++ package.json
@@ -21,7 +21,7 @@
     "test:e2e": "playwright test",
     "test:security": "node --test tests/security-logic.mjs",
     "test:monitor": "node --test tests/monitor-logic.test.mjs",
-    "test:checks": "node --test tests/security-logic.mjs tests/monitor-logic.test.mjs tests/r147-checks.test.mjs tests/r149-checks.test.mjs tests/r150-checks.test.mjs tests/r151-checks.test.mjs tests/r152-checks.test.mjs tests/r153-checks.test.mjs tests/r154-checks.test.mjs tests/r155-checks.test.mjs tests/r156-checks.test.mjs tests/r157-checks.test.mjs tests/r158-checks.test.mjs tests/r159-checks.test.mjs tests/r160-checks.test.mjs tests/r161-checks.test.mjs tests/r162-checks.test.mjs tests/r163-checks.test.mjs tests/r164-checks.test.mjs tests/r165-checks.test.mjs tests/r166-checks.test.mjs tests/r167-checks.test.mjs tests/r168-checks.test.mjs tests/r169-checks.test.mjs tests/r170-checks.test.mjs tests/r171-checks.test.mjs tests/r172-checks.test.mjs tests/r173-checks.test.mjs tests/r174-checks.test.mjs tests/r175-checks.test.mjs tests/r176-checks.test.mjs tests/r179-checks.test.mjs tests/r180-checks.test.mjs tests/r181-checks.test.mjs tests/r182-checks.test.mjs",
+    "test:checks": "node --test tests/security-logic.mjs tests/monitor-logic.test.mjs tests/r147-checks.test.mjs tests/r149-checks.test.mjs tests/r150-checks.test.mjs tests/r151-checks.test.mjs tests/r152-checks.test.mjs tests/r153-checks.test.mjs tests/r154-checks.test.mjs tests/r155-checks.test.mjs tests/r156-checks.test.mjs tests/r157-checks.test.mjs tests/r158-checks.test.mjs tests/r159-checks.test.mjs tests/r160-checks.test.mjs tests/r161-checks.test.mjs tests/r162-checks.test.mjs tests/r163-checks.test.mjs tests/r164-checks.test.mjs tests/r165-checks.test.mjs tests/r166-checks.test.mjs tests/r167-checks.test.mjs tests/r168-checks.test.mjs tests/r169-checks.test.mjs tests/r170-checks.test.mjs tests/r171-checks.test.mjs tests/r172-checks.test.mjs tests/r173-checks.test.mjs tests/r174-checks.test.mjs tests/r175-checks.test.mjs tests/r176-checks.test.mjs tests/r179-checks.test.mjs tests/r180-checks.test.mjs tests/r181-checks.test.mjs tests/r182-checks.test.mjs tests/r183-checks.test.mjs",
     "test": "node scripts/static-checks.mjs && node scripts/engine-coupling.mjs --gate && npm run test:checks && playwright test",
     "report": "playwright show-report"
   },
